Add Response media type detection (#31)
* Add custom Response type that detects media_type

The custom response type wraps the webob type with some modest changes.

* Unit test the new response type

* Example test the new Response type

* Cleanup - linting, cut out dead code, fix types

The tests were passing, but we weren't at project standards. I was
assigning None to typed variables without Optional. We had a test that
would have failed, but for the same name overwriting it later in the
file. And there were issues with the types in the attributes of the
Response where I wanted to set some properties to None at this time,
but I didn't type them as Optional.

Co-authored-by: Nathan Slaughter <nathan.slaughter@q2.com>